revert pcap move
